% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{sampleP}
\alias{sampleP}
\title{Sample person parameters}
\usage{
sampleP(Z, RT, theta_i, zeta_i, beta, lambda, theta = 0, zeta = 0, tau,
  delta, sig2, sig2k, nu, a.p = 0.5, b.p = 0.5, n0.p = 1)
}
\description{
Sample person parameters
}
